README.md
onX is home to a team of passionate explorers. We celebrate our adventures, but we don’t take them for granted. Our appreciation for wild places drives us to preserve access to the recreation that we cherish. As a company, we invest in land access initiatives to expand and protect open spaces. We believe that the more people who experience their own off-the-beaten-path awakening, the more likely they are to join us in protecting these spaces for the future.